pyproject.toml
[tool.poetry]
name = "dasi"
version = "0.0.4"
description = ""
authors = ["jvrana <justin.vrana@gmail.com>"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Development Status :: 4 - Beta"
]

[tool.poetry.dependencies]
python = "^3.6"
networkx = "^2.3"
more-itertools = "^7.1"
tqdm = "^4.32"
sortedcontainers = "^2.1"
loggable-jdv = "^0.1.5"
uvloop = "^0.12.2"
nest_asyncio = "^1.0"
msgpack = "^0.6.1"
msgpack-numpy = "^0.4.4"
sympy = "^1.4"
pandas = "^0.25.1"
numpy = "^1.17"
biopython = "^1.74"
seaborn = "^0.9.0"
matplotlib = "^3.1"
frozendict = "^1.2"
primer3plus = "^1.0.5"
pyblastbio = "^0.4.1"
fire = "^0.1"
jsonschema = "^3.1"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest = "^5.0"
pre-commit = "^1.17"
pytest-benchmark = "^3.2"
pygal = "^2.4"
pygaljs = "^1.0"
dill = "^0.3.0"
sphinx = "^2.2"
sphinx-bootstrap-theme = "^0.7.1"
nbsphinx = "^0.4.2"
recommonmark = "^0.6.0"
sphinx_autodoc_typehints = "^1.8"
flake8 = "^3.7"
autodocsumm = "^0.1.11"
jdna = "^0.1.4"
dictdiffer = "^0.8.0"
keats = "0.2.26"
sphinx-jsonschema = "^1.11"

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.poetry.scripts]
dasi = 'dasi:command_line.main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

